Use background thread to obtain full host name
wxGetFullHostName() can take up to several seconds.
Use a background thread to obtain this information
to not block the UI.

The UI can still get blocked when the values are
refreshed in a short time as we still need to wait
for the thread to finish before starting it again.